add multer for file upload in user routes for avatar by post http

+++ src/routes/task.js
@@ -23,11 +23,6 @@ router.post('/tasks', auth, async (req, res) => {
   } catch (e) {
     res.status(400).send(e);
   }
-  // task.save().then(() => {
-  //   res.status(201).send(task);
-  // }).catch((e) => {
-  //   res.status(400).send(e);
-  // });
 });
 
 //------------------READ---------------------
@@ -92,15 +87,7 @@ router.patch('/tasks/:id', auth, async (req, res) => {
   });
   // now check related are mean or not. 
   try {
-    // const task = await Task.findById(req.params.id);
-    // const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
-    //   new: true,
-    //   runValidators: true
-    // })
-    const task = await Task.findOne({
-      _id: req.params.id,
-      owner: req.user._id
-    });
+    const task = await Task.findOne({_id: req.params.id, owner: req.user._id});
     if (!task) return res.status(404).send();
     updates.forEach((update) => task[update] = req.body[update]);
     await task.save();
@@ -112,13 +99,8 @@ router.patch('/tasks/:id', auth, async (req, res) => {
 
 //------------------DELETE---------------------
 router.delete('/tasks/:id', auth, async (req, res) => {
-  // const _id = req.params.id;
-  // const task = await Task.findByIdAndDelete(_id)
   try {
-    const task = await Task.findOneAndDelete({
-      _id: req.params.id,
-      owner: req.user._id
-    })
+    const task = await Task.findOneAndDelete({_id: req.params.id,owner: req.user._id})
     if (!task) return res.status(404).send();
     res.send(task)
   } catch (e) {
